requirements.txt
openai>=2.7.2
pyyaml>=5.4.1
rich>=14.2.0
